moved ChainName from package
